+++ frontend/web/components/App.js
@@ -23,7 +23,7 @@ import GithubStar from './GithubStar'
 import Tooltip from './Tooltip'
 import classNames from 'classnames'
 import { apps, gitBranch, gitCompare, statsChart } from 'ionicons/icons'
-import NavSubLink from './NavSubLink'
+import NavSubLink from './navigation/NavSubLink'
 import SettingsIcon from './svg/SettingsIcon'
 import UsersIcon from './svg/UsersIcon'
 import BreadcrumbSeparator from './BreadcrumbSeparator'
@@ -124,16 +124,6 @@ const App = class extends Component {
     }
     this.props.history.listen(updateLastViewed)
     updateLastViewed()
-  }
-
-  toggleDarkMode = () => {
-    const newValue = !Utils.getFlagsmithHasFeature('dark_mode')
-    flagsmith.setTrait('dark_mode', newValue)
-    if (newValue) {
-      document.body.classList.add('dark')
-    } else {
-      document.body.classList.remove('dark')
-    }
   }
 
   componentDidUpdate(prevProps) {
@@ -343,6 +333,12 @@ const App = class extends Component {
     }
     const isOrganisationSelect = document.location.pathname === '/organisations'
     const integrations = Object.keys(Utils.getIntegrationData())
+    const environmentMetricsEnabled = Utils.getFlagsmithHasFeature(
+      'environment_metrics',
+    )
+    const projectMetricsTooltipEnabled = Utils.getFlagsmithHasFeature(
+      'project_metrics_tooltip',
+    )
     return (
       <Provider store={getStore()}>
         <AccountProvider
@@ -527,24 +523,6 @@ const App = class extends Component {
                                   <GithubStar />
 
                                   <Headway className='cursor-pointer' />
-                                  <Tooltip
-                                    place='bottom'
-                                    title={
-                                      <div className='dark-mode mt-0'>
-                                        <Switch
-                                          checked={Utils.getFlagsmithHasFeature(
-                                            'dark_mode',
-                                          )}
-                                          onChange={this.toggleDarkMode}
-                                          darkMode
-                                        />
-                                      </div>
-                                    }
-                                  >
-                                    {Utils.getFlagsmithHasFeature('dark_mode')
-                                      ? 'Light Mode'
-                                      : 'Dark Mode'}
-                                  </Tooltip>
                                 </Row>
                               </nav>
                             </React.Fragment>
@@ -609,6 +587,21 @@ const App = class extends Component {
                           >
                             Compare
                           </NavSubLink>
+                          {projectMetricsTooltipEnabled && (
+                            <NavSubLink
+                              icon={gitCompare}
+                              to=''
+                              id='reporting-link'
+                              disabled
+                              tooltip={
+                                Utils.getFlagsmithValue(
+                                  'project_metrics_tooltip',
+                                ) || 'Coming soon - fallback'
+                              }
+                            >
+                              Reporting
+                            </NavSubLink>
+                          )}
                           <Permission
                             level='project'
                             permission='ADMIN'
@@ -626,6 +619,27 @@ const App = class extends Component {
                               )
                             }
                           </Permission>
+                          {Utils.getFlagsmithHasFeature(
+                            'release_pipelines',
+                          ) && (
+                            <Permission
+                              level='project'
+                              permission='ADMIN'
+                              id={projectId}
+                            >
+                              {({ permission }) =>
+                                permission && (
+                                  <NavSubLink
+                                    icon={<Icon name='flash' />}
+                                    id='release-pipelines-link'
+                                    to={`/project/${projectId}/release-pipelines`}
+                                  >
+                                    Release Pipelines
+                                  </NavSubLink>
+                                )
+                              }
+                            </Permission>
+                          )}
                         </>
                       ) : (
                         !!AccountStore.getOrganisation() && (